The Coq pseudocode syntax is now well typed by Coq, but some functions still need to be defined and then maybe some semantics change.
Normalization is not done anymore in the case of SH4, because it contains an ARM specific function which sometimes translates "NOT" to "not".
If we find a way to resolve this, it can be activated later and is maybe preferable.

When I searched for an equivalent type to the "bool" one in Coq, I found "_Bool" written inside the lexer of Compcert. Then I used it in [ast.ml].

In [norm.ml], the "NOT" keyword is sometimes translated to "not", but in [gencoq.ml], some "NOT" are also translated to "not", see for example the function [fun_name].
This behaviour is correct in the case we attempt to translate the real "NOT" function but if we translate a pseudocode of the form
<<
let NOT a b... := ... in
NOT a b...
>>
we obtain a bad translation.
In [c2pc.ml], I manually alpha-replaced "NOT" by a not conflicted function name.


git-svn-id: https://scm.gforge.inria.fr/anonscm/svn/simsoc-cert@1026 ce709078-17db-4df8-89f4-d3e5bb8270d9