1) A let-simplification is done inside the body of each function "..._step". Indeed, we can simplify the constructor Let in ast.ml, this constructor only appears at the top position of each function generated from C. Finally, we do not need to calculate the global and local variable of each function using [vars] in ast.ml, because these type can be copied directly from the C code. Note that by default, each local C variable is translated in Coq as a local register Coq variable (maybe we can translate some of them to a native Coq let-definition).
Now the global variable in Coq of each function is exactly its equivalent list of parameter in C, then some wrongly parameters (such as register parameter) disappears.

2)
All specific SH4 registers is considered as a regnum instead of variable. There is no more todo in the generated Coq, except for the instruction DIV1 where we call a 'Case' construction. The generated Coq is also well-typed.


git-svn-id: https://scm.gforge.inria.fr/anonscm/svn/simsoc-cert@1043 ce709078-17db-4df8-89f4-d3e5bb8270d9