The dependent list defined in Util.v and Csyntax_print.v is replaced by the isomorphic one, 'vector', already defined in the standard library.
About the pretty-printing of [positive] and [Z], it is now done at Coq side instead of OCaml. Note that the previous one I have done is not complete because for example, the [_float] function make a call to [_Z] and only [_Z] and [_positive] have been modified...
Now, we can choose if we want to expand the number or have a simplified shape, by instanciating what we want at line (Csyntax_print.v, 691 and 692).
Remark that I have given 2 definitions of [string_of_nat_aux], there is one which is not accepted by Coq 8.2. When we will use 8.3, we can replace it.


git-svn-id: https://scm.gforge.inria.fr/anonscm/svn/simsoc-cert@1081 ce709078-17db-4df8-89f4-d3e5bb8270d9